Moved testing of loading of non-emmo based ontologies to separate test
